코드 정리, LoadVideo 변경
사용하지 않는 코드 제거
LoadVideo 대신 생성자에 비디오 주소를 넣게 변경
